Treat commit hook execution in observed repositories as a no-op, not an error
Summary:
See PHI24. If you create a hosted Mercurial repository and switch it to observed, you can end up with a hook installed that runs on pulls and complains.

Instead, just bail out if we're running on a pull.

The corresponding Git hook doesn't run on pulls, so there's no issue in Git.

Test Plan: Executed the hook in an observed Mercurial repository, got a clean exit.

Reviewers: chad

Reviewed By: chad

Differential Revision: https://secure.phabricator.com/D18307