Add custom field 'components for review' to arc diff
Summary:
This commit is looking to add a new custom field to diffs so that the author of the diff can define the components whose fixtures they want to be made available for review. Components defined here should be listed in a comma-separated list, though this is an optional field. When included, these components will be used to craft the fixture links for the corresponding znd deploy of the phabricator diff, making it available for review in react-sandbox on the live site.

NOTE: This feature is not yet fully implemented on the phabricator/jenkins side, and it remains unconfirmed whether or not Phabricator still supports custom fields for diffs. So this commit is for testing purposes and will require followup documentation when/if we get it working.

Relevant JIRA ticket: https://khanacademy.atlassian.net/browse/JSBUILD-194

Test Plan: Fingers crossed

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D41547